Use the same tooltip style for pie charts.
